+++ reconstruction/src/world.rs
@@ -12,19 +12,19 @@ pub struct World {
     /// Name of original domain
     domain_name: String,
     /// Maps type name to its index
-    types: HashMap<String, u32>,
+    types: HashMap<String, u16>,
     /// The parent of each index
-    type_children: HashMap<u32, Vec<u32>>,
+    type_children: HashMap<u16, Vec<u16>>,
     /// Maps predicate name to its index
-    predicates: HashMap<String, u32>,
+    predicates: HashMap<String, u16>,
     /// Maps action name to its index
-    actions: HashMap<String, u32>,
+    actions: HashMap<String, u16>,
     /// Maps meta action name to its index
-    meta_actions: HashMap<String, u32>,
+    meta_actions: HashMap<String, u16>,
     /// Maps object name to its index
-    objects: HashMap<String, u32>,
+    objects: HashMap<String, u16>,
     /// Maps object index to its type index
-    object_types: HashMap<u32, u32>,
+    object_types: HashMap<u16, u16>,
 }
 
 pub static WORLD: OnceCell<World> = OnceCell::new();
@@ -61,11 +61,11 @@ impl World {
         &self.domain_name
     }
 
-    pub fn get_default_type_index(&self) -> u32 {
+    pub fn get_default_type_index(&self) -> u16 {
         0
     }
 
-    pub fn get_type_index(&self, name: &str) -> u32 {
+    pub fn get_type_index(&self, name: &str) -> u16 {
         assert!(
             self.types.contains_key(name),
             "Found undeclared type: {}. Actual types: {:?}",
@@ -75,15 +75,15 @@ impl World {
         self.types[name]
     }
 
-    pub fn get_type_name(&self, index: u32) -> &String {
+    pub fn get_type_name(&self, index: u16) -> &String {
         &self.types.iter().find(|(_, i)| **i == index).unwrap().0
     }
 
-    pub fn get_type_children(&self, index: u32) -> Option<&Vec<u32>> {
+    pub fn get_type_children(&self, index: u16) -> Option<&Vec<u16>> {
         self.type_children.get(&index)
     }
 
-    pub fn is_of_type(&self, type_index: u32, wished_index: u32) -> bool {
+    pub fn is_of_type(&self, type_index: u16, wished_index: u16) -> bool {
         if type_index == wished_index {
             return true;
         }
@@ -94,11 +94,11 @@ impl World {
         }
     }
 
-    pub fn get_action_index(&self, name: &str) -> u32 {
+    pub fn get_action_index(&self, name: &str) -> u16 {
         self.actions[name]
     }
 
-    pub fn get_action_name(&self, index: u32) -> &String {
+    pub fn get_action_name(&self, index: u16) -> &String {
         &self.actions.iter().find(|(_, i)| **i == index).unwrap().0
     }
 
@@ -106,11 +106,11 @@ impl World {
         self.meta_actions.contains_key(name) && !self.actions.contains_key(name)
     }
 
-    pub fn get_meta_index(&self, name: &str) -> u32 {
+    pub fn get_meta_index(&self, name: &str) -> u16 {
         self.meta_actions[name]
     }
 
-    pub fn get_meta_name(&self, index: u32) -> &String {
+    pub fn get_meta_name(&self, index: u16) -> &String {
         &self
             .meta_actions
             .iter()
@@ -119,11 +119,11 @@ impl World {
             .0
     }
 
-    pub fn get_predicate_index(&self, name: &str) -> u32 {
+    pub fn get_predicate_index(&self, name: &str) -> u16 {
         self.predicates[name]
     }
 
-    pub fn get_predicate_name(&self, index: u32) -> &String {
+    pub fn get_predicate_name(&self, index: u16) -> &String {
         &self
             .predicates
             .iter()
@@ -132,38 +132,38 @@ impl World {
             .0
     }
 
-    pub fn get_object_count(&self) -> u32 {
-        self.objects.len() as u32
+    pub fn get_object_count(&self) -> u16 {
+        self.objects.len() as u16
     }
 
-    pub fn get_object_index(&self, name: &str) -> u32 {
+    pub fn get_object_index(&self, name: &str) -> u16 {
         self.objects[name]
     }
 
-    pub fn get_object_indexes(&self, indexes: &Vec<String>) -> Vec<u32> {
+    pub fn get_object_indexes(&self, indexes: &Vec<String>) -> Vec<u16> {
         indexes.iter().map(|i| self.get_object_index(i)).collect()
     }
 
-    pub fn get_object_name(&self, index: u32) -> &String {
+    pub fn get_object_name(&self, index: u16) -> &String {
         &self.objects.iter().find(|(_, i)| **i == index).unwrap().0
     }
 
-    pub fn get_object_names(&self, indexes: &Vec<u32>) -> Vec<&String> {
+    pub fn get_object_names(&self, indexes: &Vec<u16>) -> Vec<&String> {
         indexes.iter().map(|i| self.get_object_name(*i)).collect()
     }
 
-    pub fn get_object_names_cloned(&self, indexes: &Vec<u32>) -> Vec<String> {
+    pub fn get_object_names_cloned(&self, indexes: &Vec<u16>) -> Vec<String> {
         indexes
             .iter()
             .map(|i| self.get_object_name(*i).to_owned())
             .collect()
     }
 
-    pub fn get_object_type(&self, object: u32) -> u32 {
+    pub fn get_object_type(&self, object: u16) -> u16 {
         self.object_types[&object]
     }
 
-    pub fn iterate_objects<'a>(&'a self) -> impl Iterator<Item = (u32, u32)> + 'a {
+    pub fn iterate_objects<'a>(&'a self) -> impl Iterator<Item = (u16, u16)> + 'a {
         (0..self.get_object_count())
             .into_iter()
             .map(|i| (i, self.get_object_type(i)))
@@ -177,7 +177,7 @@ impl World {
         })
     }
 
-    pub fn get_objects_with_type(&self, type_id: u32) -> Vec<u32> {
+    pub fn get_objects_with_type(&self, type_id: u16) -> Vec<u16> {
         World::global()
             .iterate_objects()
             .filter_map(
@@ -190,20 +190,20 @@ impl World {
     }
 }
 
-fn extract_types(types: &Option<Types>) -> (HashMap<String, u32>, HashMap<u32, Vec<u32>>) {
-    let mut index_map: HashMap<String, u32> = HashMap::new();
-    let mut type_children: HashMap<u32, Vec<u32>> = HashMap::new();
+fn extract_types(types: &Option<Types>) -> (HashMap<String, u16>, HashMap<u16, Vec<u16>>) {
+    let mut index_map: HashMap<String, u16> = HashMap::new();
+    let mut type_children: HashMap<u16, Vec<u16>> = HashMap::new();
     index_map.insert("object".to_string(), 0);
 
     if let Some(types) = types {
         for t in types.iter() {
             if !index_map.contains_key(&t.name) {
-                index_map.insert(t.name.to_owned(), index_map.len() as u32);
+                index_map.insert(t.name.to_owned(), index_map.len() as u16);
             }
             let type_index = index_map.get(&t.name).unwrap().to_owned();
             for t in t.sub_types.iter() {
                 if !index_map.contains_key(t) {
-                    index_map.insert(t.to_owned(), index_map.len() as u32);
+                    index_map.insert(t.to_owned(), index_map.len() as u16);
                 }
                 let child_index = index_map.get(t).unwrap();
                 let children_entry = &mut type_children.entry(type_index).or_default();
@@ -215,45 +215,45 @@ fn extract_types(types: &Option<Types>) -> (HashMap<String, u32>, HashMap<u32, V
     (index_map, type_children)
 }
 
-fn extract_predicates(predicates: &Predicates) -> HashMap<String, u32> {
+fn extract_predicates(predicates: &Predicates) -> HashMap<String, u16> {
     predicates
         .iter()
         .enumerate()
-        .map(|(i, p)| (p.name.to_owned(), i as u32))
+        .map(|(i, p)| (p.name.to_owned(), i as u16))
         .collect()
 }
 
-fn extract_actions(actions: &Actions) -> HashMap<String, u32> {
+fn extract_actions(actions: &Actions) -> HashMap<String, u16> {
     actions
         .iter()
         .enumerate()
-        .map(|(i, a)| (a.name.to_owned(), i as u32))
+        .map(|(i, a)| (a.name.to_owned(), i as u16))
         .collect()
 }
 
 fn extract_meta_actions(
-    actions: &HashMap<String, u32>,
+    actions: &HashMap<String, u16>,
     meta_actions: &Actions,
-) -> HashMap<String, u32> {
-    let mut index_map: HashMap<String, u32> = HashMap::new();
+) -> HashMap<String, u16> {
+    let mut index_map: HashMap<String, u16> = HashMap::new();
     for (i, a) in meta_actions.iter().enumerate() {
         if !actions.contains_key(&a.name) {
-            index_map.insert(a.name.to_owned(), i as u32);
+            index_map.insert(a.name.to_owned(), i as u16);
         }
     }
     index_map
 }
 
 fn extract_objects(
-    type_map: &HashMap<String, u32>,
+    type_map: &HashMap<String, u16>,
     objects: &Objects,
-) -> (HashMap<String, u32>, HashMap<u32, u32>) {
-    let temp: Vec<((String, u32), (u32, u32))> = objects
+) -> (HashMap<String, u16>, HashMap<u16, u16>) {
+    let temp: Vec<((String, u16), (u16, u16))> = objects
         .iter()
         .enumerate()
         .map(|(i, o)| {
             let object_name = o.name.to_owned();
-            let object_index = i as u32;
+            let object_index = i as u16;
             let object_type = match &o.type_name {
                 Some(t) => t,
                 None => "object",
